requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
marqo==2.1.0
packaging==23.2
pydantic==1.10.13
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.1.0
